--- trisonics-scouting/src/app/components/garv/garv.component.html
@@ -1,1 +1,0 @@
-<p>garv works!</p>
