--- trisonics-scouting/src/app/components/garv/garv.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-garv',
-  templateUrl: './garv.component.html',
-  styleUrls: ['./garv.component.scss']
-})
-export class GarvComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
