+++ trisonics-scouting/src/app/components/pit-view/pit-view.component.scss
@@ -1,0 +1,11 @@
+.image-container {
+  width: 300px;
+  margin-top: 1em;
+  margin-bottom: 1em;
+}
+
+.notes-container {
+  width: 300px;
+  margin-top: 1em;
+  margin-bottom: 1em;
+}
