+++ trisonics-scouting/src/app/components/scout-match-view/scout-match-view.component.html
@@ -223,8 +223,8 @@
   </table>
   <div class="flex-col">
     <div *ngFor="let pr of this.pitData" class="pit-card">
-      <app-scout-pit-view [pitResult]="pr">
-      </app-scout-pit-view>
+      <app-pit-view [pitResult]="pr">
+      </app-pit-view>
     </div>
   </div>
 </ng-container>
