+++ trisonics-scouting/src/app/components/scout-match-view/scout-match-view.component.scss
