+++ trisonics-scouting/src/app/components/scout-match/scout-match.component.scss
