--- trisonics-scouting/src/app/components/scout-pit-view/scout-pit-view.component.scss
@@ -1,11 +1,0 @@
-.image-container {
-  width: 300px;
-  margin-top: 1em;
-  margin-bottom: 1em;
-}
-
-.notes-container {
-  width: 300px;
-  margin-top: 1em;
-  margin-bottom: 1em;
-}
