+++ trisonics-scouting/src/app/components/scout-pit/scout-pit.component.html
@@ -96,8 +96,8 @@
     </p>
     <div class="flex-row">
       <div *ngFor="let pr of this.pitResultList">
-        <app-scout-pit-view [pitResult]="pr" [matches]="false">
-        </app-scout-pit-view>
+        <app-pit-view [pitResult]="pr" [matches]="false">
+        </app-pit-view>
       </div>
     </div>
   </div>
