--- trisonics-scouting/src/app/components/view-pits/view-pits.component.html
@@ -1,20 +1,0 @@
-<div class="flex-row" *ngIf="this.teamsLoaded && this.resultsLoaded">
-    <ng-container *ngFor="let t of this.teamList">
-        <div *ngFor="let pr of getPits(t.number)">
-            <app-scout-pit-view [pitResult]="pr" [matches]="true">
-            </app-scout-pit-view>
-        </div>
-        <div *ngIf="getPits(t.number).length === 0">
-            <mat-card>
-                <mat-card-title>
-                    {{t.number}} - {{t.name}}
-                </mat-card-title>
-                <mat-card-subtitle>
-                    <strong>No data scouted yet</strong>
-                </mat-card-subtitle>
-                <mat-card-content>
-                </mat-card-content>
-            </mat-card>
-        </div>
-    </ng-container>
-</div>
